Launchpad.Web/slither-api/requirements.txt
Flask==2.3.3
py-solc-x==2.0.4
slither-analyzer==0.10.0
setuptools>=65.0.0  # Required for pkg_resources

# Optional dependencies for production
gunicorn==21.2.0  # WSGI server for production
flask-cors==4.0.0  # Enable CORS if needed for frontend integration
python-dotenv==1.0.0  # Environment variable management

# Optional dependencies for development
pytest==7.4.3  # Testing framework
pytest-flask==1.3.0  # Flask-specific testing utilities